+++ src/types.ts
@@ -1,5 +1,5 @@
 import { Document } from 'mongodb'
 
 export interface SyncOptions {
-  mapper: (doc: Document) => Document
+  mapper?: (doc: Document) => Document
 }